requirements.txt
requests
python-telegram-bot
sparqlwrapper
